refactor(design-system): move the QueryFilter translation into the app
The design system had no business knowing the backend request shape: nothing in
the package used `routeQueryToQueryFilters` (only `index.ts` re-exported it) and
every caller is app-side, yet building the payload dragged
`@kestra-io/kestra-sdk` in as a peer dependency of a package published for its
components.

Split it along the ownership line instead. The design system keeps the URL key
format it writes via `encodeFiltersToQuery` and now exposes `parseFilterKey` +
`decodeFilterValue` for it, so the regex still lives in exactly one place and
`decodeSearchParams` reads through the same parser. The translation to
`QueryFilter[]` moves to `ui/src/utils/queryFilters.ts`, typed against the SDK
there, and the peer dependency is gone.

Callers now import it from the app util, and `logs.ts` drops its
`as QueryFilter[]` cast since the util is typed. Tests move alongside the code,
with new coverage for `parseFilterKey`.
