+++ ui/packages/design-system/src/components/Data/KsDataTable/filter/utils/helpers.ts
@@ -1,11 +1,12 @@
 import type {LocationQuery} from "vue-router"
 import {type AppliedFilter, type FilterGroup, type LeafFilterGroup, type LogicalOperator, Comparators, isWrapperGroup} from "./filterTypes"
 import {MAX_RENDERABLE_NESTING_DEPTH} from "./constants"
-import type {QueryFilter, QueryFilterField, QueryFilterLogical, QueryFilterOp} from "@kestra-io/kestra-sdk"
 
-export type {QueryFilter}
-
-const decodeURIComponentSafely = (value: string | (string | null)[]): string | string[] =>
+/**
+ * Decodes a raw `filters[...]` query-param value. Exported because callers translating the route
+ * into a backend request payload decode the same values as {@link decodeSearchParams} does.
+ */
+export const decodeFilterValue = (value: string | (string | null)[]): string | string[] =>
     Array.isArray(value)
         ? value.filter(v => v !== null).map(decodeURIComponent)
         : decodeURIComponent(value)
@@ -33,7 +34,7 @@ const FILTER_KEY_PATTERN = new RegExp(
 
 const PREFIX_SEGMENT_PATTERN = /\[(and|or)]\[(\d+)]/gi
 
-interface PrefixSegment {
+export interface PrefixSegment {
     logical: LogicalOperator
     index: number
 }
@@ -59,103 +60,27 @@ export interface DecodedParam {
     wrapperLogical?: LogicalOperator
 }
 
-
-const QUERY_FILTER_LOGICAL: Record<LogicalOperator, QueryFilterLogical> = {AND: "and", OR: "or"}
-
-/**
- * Builds a `QueryFilter` group node. The backend rejects a node that also carries a
- * `field`/`operation`/`value` (see `QueryFilter`'s canonical constructor), so a group is only ever
- * `logical` + `children`, with `logical` serialized in the lowercase form its enum accepts.
- */
-const queryFilterNode = (logical: LogicalOperator, children: QueryFilter[]): QueryFilter =>
-    ({logical: QUERY_FILTER_LOGICAL[logical], children})
-
-/**
- * Builds one logical position in the filter tree, mirroring the backend's
- * `QueryFilterFormatBinder.NodeBuilder`: direct leaves land here directly, LABELS leaves sharing an
- * operation are merged into one filter with a map value, and nested `[and|or][N]` segments descend
- * into a sub-node keyed by (logical, index) that is recursively flattened by {@link build}.
- */
-class FilterNodeBuilder {
-    private readonly directLeaves: QueryFilter[] = []
-    private readonly labelsByOp = new Map<QueryFilterOp, Record<string, string>>()
-    private readonly subNodes = new Map<LogicalOperator, Map<number, FilterNodeBuilder>>()
-
-    descend(logical: LogicalOperator, index: number): FilterNodeBuilder {
-        const slots = this.subNodes.get(logical) ?? new Map<number, FilterNodeBuilder>()
-        this.subNodes.set(logical, slots)
-        const slot = slots.get(index) ?? new FilterNodeBuilder()
-        slots.set(index, slot)
-        return slot
-    }
-
-    addLeaf(field: QueryFilterField, operation: QueryFilterOp, subKey: string | undefined, value: string | string[]) {
-        const scalarValue = Array.isArray(value) ? value[0] : value
-
-        if (field === "labels" && subKey) {
-            const map = this.labelsByOp.get(operation) ?? {}
-            map[subKey] = scalarValue
-            this.labelsByOp.set(operation, map)
-            return
-        }
-
-        this.directLeaves.push({field, operation, value})
-    }
-
-    build(): QueryFilter[] {
-        const items: QueryFilter[] = [...this.directLeaves]
-
-        this.labelsByOp.forEach((map, operation) => {
-            if (Object.keys(map).length > 0) {
-                items.push({field: "labels", operation, value: map})
-            }
-        })
-
-        this.subNodes.forEach((slots, logical) => {
-            const branches: QueryFilter[] = []
-            slots.forEach((slot) => {
-                const slotItems = slot.build()
-                if (slotItems.length === 0) return
-                branches.push(slotItems.length === 1
-                    ? slotItems[0]
-                    : queryFilterNode("AND", slotItems))
-            })
-            if (branches.length === 0) return
-            items.push(branches.length === 1 && logical === "AND"
-                ? branches[0]
-                : queryFilterNode(logical, branches))
-        })
-
-        return items
-    }
+/** One `filters[and|or][N]…[field][OPERATION][subKey]` key, split into its parts. */
+export interface ParsedFilterKey {
+    /** The `[and|or][N]` grouping chain, outermost first; empty for a root-level filter. */
+    chain: PrefixSegment[]
+    field: string
+    operation: string
+    /** Present for keys carrying a sub-key, e.g. the label name in `filters[labels][EQUALS][env]`. */
+    subKey?: string
 }
 
 /**
- * Converts a route's raw `filters[...]` query params (as produced by {@link encodeFiltersToQuery} /
- * {@link encodeFilterGroupsToQuery}) directly into the `QueryFilter[]` array shape the generated
- * `@kestra-io/kestra-sdk` search functions expect. Unlike {@link decodeSearchParams} — which targets
- * restoring `KsFilter` UI chips and lossily flattens a labels `subKey` into a `"key:value"` string —
- * this preserves the LABELS sub-key structure and nested AND/OR grouping needed for a faithful
- * backend request, by porting `QueryFilterFormatBinder`'s regex + tree-building.
+ * Parses a filter URL key into its parts, or returns null when the key does not match
+ * {@link FILTER_KEY_PATTERN}. Exported so callers that translate the route into a backend request
+ * payload read the key format from its owner instead of restating the regex.
  */
-export const routeQueryToQueryFilters = (query: LocationQuery): QueryFilter[] => {
-    const root = new FilterNodeBuilder()
+export const parseFilterKey = (key: string): ParsedFilterKey | null => {
+    const match = key.match(FILTER_KEY_PATTERN)
+    if (!match) return null
 
-    for (const [key, value] of Object.entries(query)) {
-        if (!key.startsWith("filters[") || !value) continue
-        const match = key.match(FILTER_KEY_PATTERN)
-        if (!match) continue
-
-        const [, prefix, field, operation, subKey] = match
-        const chain = parsePrefixChain(prefix)
-        let target = root
-        for (const segment of chain) {
-            target = target.descend(segment.logical, segment.index)
-        }
-        target.addLeaf(field as QueryFilterField, operation as QueryFilterOp, subKey, decodeURIComponentSafely(value) as string | string[])
-    }
-
-    return root.build()
+    const [, prefix, field, operation, subKey] = match
+    return {chain: parsePrefixChain(prefix), field, operation, subKey}
 }
 
 export const decodeSearchParams = (query: LocationQuery): DecodedParam[] =>
@@ -163,12 +88,10 @@ export const decodeSearchParams = (query: LocationQuery): DecodedParam[] =>
         .filter(([key]) => key.startsWith("filters[") || key === "q")
         .map(([key, value]): DecodedParam | null => {
             if (!value) return null
-            const match = key.match(FILTER_KEY_PATTERN)
-            if (!match) return null
+            const parsed = parseFilterKey(key)
+            if (!parsed) return null
 
-            const [, prefix, field, operation, subKey] = match
-            const chain = parsePrefixChain(prefix)
-            return buildParam(field, operation, subKey, value, chain)
+            return buildParam(parsed.field, parsed.operation, parsed.subKey, value, parsed.chain)
         })
         .filter((v): v is DecodedParam => v !== null)
 
@@ -180,8 +103,8 @@ const buildParam = (
     chain: PrefixSegment[],
 ): DecodedParam => {
     const decoded = subKey
-        ? `${subKey}:${decodeURIComponentSafely(value)}`
-        : decodeURIComponentSafely(value)
+        ? `${subKey}:${decodeFilterValue(value)}`
+        : decodeFilterValue(value)
     return {
         field,
         value: decoded,
