+++ ui/packages/design-system/src/index.ts
@@ -173,8 +173,10 @@ export {
     serializeFiltersToString,
     parseFiltersFromString,
     validStructureSignature,
-    routeQueryToQueryFilters,
-    type QueryFilter,
+    parseFilterKey,
+    decodeFilterValue,
+    type ParsedFilterKey,
+    type PrefixSegment,
 } from "./components/Data/KsDataTable/filter/utils/helpers"
 export {pickStarterField} from "./components/Data/KsDataTable/filter/utils/filterChipFactory"
 export {
